requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pymupdf4llm==0.3.4
markdown==3.9
